refactor:move loadbalancer to discovery module.
