+++ pom.xml
@@ -41,7 +41,6 @@
 		<module>spring-cloud-tencent-commons</module>
 		<module>spring-cloud-tencent-polaris-context</module>
 		<module>spring-cloud-tencent-rpc-enhancement</module>
-		<module>spring-cloud-tencent-polaris-loadbalancer</module>
 		<module>spring-cloud-starter-tencent-metadata-transfer</module>
 		<module>spring-cloud-starter-tencent-polaris-config</module>
 		<module>spring-cloud-starter-tencent-polaris-discovery</module>
