+++ spring-cloud-starter-tencent-polaris-router/pom.xml
@@ -17,10 +17,6 @@
 		<!-- Spring Cloud Tencent dependencies start -->
 		<dependency>
 			<groupId>com.tencent.cloud</groupId>
-			<artifactId>spring-cloud-tencent-polaris-loadbalancer</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>com.tencent.cloud</groupId>
 			<artifactId>spring-cloud-tencent-rpc-enhancement</artifactId>
 		</dependency>
 		<dependency>
