+++ spring-cloud-tencent-dependencies/pom.xml
@@ -119,12 +119,6 @@
 
 			<dependency>
 				<groupId>com.tencent.cloud</groupId>
-				<artifactId>spring-cloud-tencent-polaris-loadbalancer</artifactId>
-				<version>${revision}</version>
-			</dependency>
-
-			<dependency>
-				<groupId>com.tencent.cloud</groupId>
 				<artifactId>spring-cloud-starter-tencent-metadata-transfer</artifactId>
 				<version>${revision}</version>
 			</dependency>
