+++ spring-cloud-tencent-examples/polaris-circuitbreaker-example/polaris-circuitbreaker-example-a/src/main/java/com/tencent/cloud/polaris/circuitbreaker/example/ServiceAController.java
@@ -17,8 +17,6 @@
 
 package com.tencent.cloud.polaris.circuitbreaker.example;
 
-import org.owasp.esapi.ESAPI;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -64,12 +62,6 @@ public class ServiceAController {
 		ResponseEntity<String> entity = restTemplate.getForEntity(
 				"http://polaris-circuitbreaker-example-b/example/service/b/info",
 				String.class);
-		String response = entity.getBody();
-		return cleanXSS(response);
-	}
-
-	private String cleanXSS(String str) {
-		str = ESAPI.encoder().encodeForHTML(str);
-		return str;
+		return entity.getBody();
 	}
 }
