+++ spring-cloud-tencent-examples/polaris-discovery-example/discovery-caller-service/pom.xml
@@ -41,11 +41,6 @@
 		<!--            <groupId>org.springframework.cloud</groupId>-->
 		<!--            <artifactId>spring-cloud-starter-netflix-eureka-client</artifactId>-->
 		<!--        </dependency>-->
-
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-loadbalancer</artifactId>
-		</dependency>
 	</dependencies>
 
 	<build>
