+++ spring-cloud-tencent-examples/polaris-gateway-example/gateway-callee-service/pom.xml
@@ -28,10 +28,5 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
-
-		<dependency>
-			<groupId>org.owasp.esapi</groupId>
-			<artifactId>esapi</artifactId>
-		</dependency>
 	</dependencies>
 </project>
