+++ spring-cloud-tencent-examples/polaris-gateway-example/gateway-scg-service/pom.xml
@@ -48,10 +48,5 @@
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-gateway</artifactId>
 		</dependency>
-
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-loadbalancer</artifactId>
-		</dependency>
 	</dependencies>
 </project>
