+++ spring-cloud-tencent-examples/polaris-router-example/router-callee-service1/pom.xml
@@ -27,11 +27,6 @@
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-bootstrap</artifactId>
 		</dependency>
-
-		<dependency>
-			<groupId>org.owasp.esapi</groupId>
-			<artifactId>esapi</artifactId>
-		</dependency>
 	</dependencies>
 
 	<build>
