--- spring-cloud-tencent-examples/polaris-router-example/router-callee-service1/src/main/resources/ESAPI.properties
@@ -1,14 +1,0 @@
-ESAPI.printProperties=true
-ESAPI.Encoder=org.owasp.esapi.reference.DefaultEncoder
-ESAPI.Logger=org.owasp.esapi.logging.slf4j.Slf4JLogFactory
-
-Encoder.AllowMultipleEncoding=false
-Encoder.AllowMixedEncoding=false
-Encoder.DefaultCodecList=HTMLEntityCodec,PercentCodec,JavaScriptCodec
-
-Logger.LogEncodingRequired=false
-Logger.UserInfo=false
-Logger.ClientInfo=false
-Logger.ApplicationName=ExampleApplication
-Logger.LogApplicationName=false
-Logger.LogServerIP=false
