+++ spring-cloud-tencent-examples/polaris-router-example/router-callee-service2/src/main/java/com/tencent/cloud/polaris/router/example/RouterCalleeController.java
@@ -18,7 +18,6 @@
 
 package com.tencent.cloud.polaris.router.example;
 
-import org.owasp.esapi.ESAPI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -50,14 +49,6 @@ public class RouterCalleeController {
 	@PostMapping("/info")
 	public String info(@RequestParam("name") String name, @RequestBody User user) {
 		LOG.info("Discovery Service Callee [{}] is called.", port);
-		return String.format("Discovery Service Callee [%s] is called. user = %s", port, cleanXSS(user));
-	}
-
-	private User cleanXSS(User user) {
-		User u = new User();
-		String name = ESAPI.encoder().encodeForHTML(user.getName());
-		u.setName(name);
-		u.setAge(user.getAge());
-		return u;
+		return String.format("Discovery Service Callee [%s] is called. user = %s", port, user);
 	}
 }
