+++ spring-cloud-tencent-examples/pom.xml
@@ -32,15 +32,4 @@
 	<properties>
 		<maven.deploy.skip>true</maven.deploy.skip>
 	</properties>
-
-	<dependencyManagement>
-		<dependencies>
-			<dependency>
-				<groupId>org.owasp.esapi</groupId>
-				<artifactId>esapi</artifactId>
-				<version>2.5.0.0</version>
-			</dependency>
-		</dependencies>
-	</dependencyManagement>
-
 </project>
