+++ spring-cloud-tencent-polaris-context/pom.xml
@@ -91,6 +91,11 @@
 
 		<dependency>
 			<groupId>com.tencent.polaris</groupId>
+			<artifactId>polaris-router-factory</artifactId>
+		</dependency>
+
+		<dependency>
+			<groupId>com.tencent.polaris</groupId>
 			<artifactId>loadbalancer-random</artifactId>
 		</dependency>
