--- spring-cloud-tencent-polaris-loadbalancer/pom.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xmlns="http://maven.apache.org/POM/4.0.0"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<parent>
-		<artifactId>spring-cloud-tencent</artifactId>
-		<groupId>com.tencent.cloud</groupId>
-		<version>${revision}</version>
-		<relativePath>../pom.xml</relativePath>
-	</parent>
-	<modelVersion>4.0.0</modelVersion>
-
-	<artifactId>spring-cloud-tencent-polaris-loadbalancer</artifactId>
-	<name>Spring Cloud Tencent Polaris LoadBalancer</name>
-
-	<dependencies>
-		<!-- Spring Cloud Tencent dependencies start -->
-		<dependency>
-			<groupId>com.tencent.cloud</groupId>
-			<artifactId>spring-cloud-tencent-polaris-context</artifactId>
-		</dependency>
-		<!-- Spring Cloud Tencent dependencies end -->
-
-		<!-- Polaris dependencies start -->
-		<dependency>
-			<groupId>com.tencent.polaris</groupId>
-			<artifactId>polaris-router-factory</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>com.tencent.polaris</groupId>
-			<artifactId>polaris-discovery-api</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>com.tencent.polaris</groupId>
-			<artifactId>polaris-test-common</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<!-- Polaris dependencies end -->
-
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-loadbalancer</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-test</artifactId>
-			<scope>test</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>org.mockito</groupId>
-			<artifactId>mockito-inline</artifactId>
-			<scope>test</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>org.mockito</groupId>
-			<artifactId>mockito-core</artifactId>
-			<scope>test</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>net.bytebuddy</groupId>
-			<artifactId>byte-buddy</artifactId>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
-
-</project>
