--- spring-cloud-tencent-polaris-loadbalancer/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,22 +1,0 @@
-{
-  "properties": [
-    {
-      "name": "spring.cloud.polaris.loadbalancer.enabled",
-      "type": "java.lang.Boolean",
-      "defaultValue": "true",
-      "description": "polaris loadbalancer."
-    },
-    {
-      "name": "spring.cloud.polaris.loadbalancer.discoveryType",
-      "type": "java.lang.String",
-      "defaultValue": "POLARIS",
-      "description": "Type of discovery server."
-    },
-    {
-      "name": "spring.cloud.polaris.loadbalancer.strategy",
-      "type": "java.lang.String",
-      "defaultValue": "random",
-      "description": "retry,best_available,availability_filtering,round_robin,weighted_response_time,zone_avoidance,random,consistent_hash,weighted_random."
-    }
-  ]
-}
