--- spring-cloud-tencent-polaris-loadbalancer/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.tencent.cloud.polaris.loadbalancer.config.PolarisLoadBalancerAutoConfiguration
